+++ test/NuGet.Clients.Tests/NuGet.CommandLine.Test/NuGetRestoreCommandTest.cs
@@ -3338,61 +3338,6 @@ EndProject";
         }
 
         [SkipMono()]
-        public void InstallCommand_WithPackageWithVulnerabilities_RaisesWarnings()
-        {
-            // Arrange
-            var nugetexe = Util.GetNuGetExePath();
-            using var pathContext = new SimpleTestPathContext();
-            using var mockServer = new FileSystemBackedV3MockServer(pathContext.PackageSource, sourceReportsVulnerabilities: true);
-
-            mockServer.Vulnerabilities.Add(
-                "packageA",
-                new List<(Uri, PackageVulnerabilitySeverity, VersionRange)> {
-            (new Uri("https://contoso.com/advisories/12345"), PackageVulnerabilitySeverity.High, VersionRange.Parse("[1.0.0, 2.0.0)"))
-                });
-
-            pathContext.Settings.RemoveSource("source");
-            pathContext.Settings.AddSource("source", mockServer.ServiceIndexUri, allowInsecureConnectionsValue: "True");
-
-            Util.CreateTestPackage("packageA", "1.1.0", pathContext.PackageSource);
-            Util.CreateTestPackage("packageB", "2.2.0", pathContext.PackageSource);
-
-            var solution = new SimpleTestSolutionContext(pathContext.SolutionRoot);
-            var projectA = new SimpleTestProjectContext(
-                "a",
-                ProjectStyle.PackagesConfig,
-                pathContext.SolutionRoot);
-
-            solution.Projects.Add(projectA);
-            solution.Create(pathContext.SolutionRoot);
-
-            Util.CreateFile(Path.GetDirectoryName(projectA.ProjectPath), "packages.config",
-@"<packages>
-  <package id=""packageA"" version=""1.1.0"" />
-  <package id=""packageB"" version=""2.2.0"" />
-</packages>");
-
-            mockServer.Start();
-
-            var installDirectory = Path.Combine(pathContext.WorkingDirectory, "install_output");
-
-            // Act
-            var r = CommandRunner.Run(
-                nugetexe,
-                pathContext.WorkingDirectory,
-                $"install {Path.Combine(Path.GetDirectoryName(projectA.ProjectPath), "packages.config")} -OutputDirectory {installDirectory}");
-
-            mockServer.Stop();
-
-            // Assert
-            r.Success.Should().BeTrue(because: r.AllOutput);
-            var installedPackage = Path.Combine(installDirectory, "packageA.1.1.0", "packageA.1.1.0.nupkg");
-            File.Exists(installedPackage).Should().BeTrue();
-            r.AllOutput.Should().Contain("Package 'packageA' 1.1.0 has a known high severity vulnerability");
-            r.AllOutput.Should().NotContain("Package 'packageB' 2.2.0 has a known high severity vulnerability");
-        }
-
-        [SkipMono()]
         public void RestoreCommand_WithProjectWithPackagesConfig_WithNuGetAuditFalse_PackageWithVulnerabilities_DoesNotRaiseWarnings()
         {
             // Arrange
